homework 7
added desctruturizing + splitted js code into modules and added some array functions
