+++ element_creation.js
@@ -1,0 +1,7 @@
+export function createElement(tag, className) {
+    const $tag = document.createElement(tag);
+    if(className) {
+        $tag.classList.add(className);
+    }
+    return $tag;
+}
