+++ random_number.js
@@ -1,0 +1,4 @@
+export function getRandomNumber(max) {
+    const randomNumber = Math.ceil(Math.random() * max);
+    return randomNumber;
+}
