+++ selectors.js
@@ -1,0 +1,4 @@
+export const $arena = document.querySelector('.arenas');
+export const $fightButton = document.querySelector('.button');
+export const $formFight = document.querySelector('.control');
+export const $chat = document.querySelector('.chat');
